Handle changes in remat position
If remat is applied to other layers, the name of the parameters will change (they will have a Checkpoint- prefix). This will break checkpoint loading. A (hacky) way to solve this is to overwrite the variable names of the checkpoint with the variable names that we have in self.state.